src/backend/requirements.txt
Flask
pymongo
bson.objectid